Respawn Mechanic Added
A simple respawn mechanic has been added via scene switching. A new scene has been added with a fully black rawImage and basic script. The script switches the black empty scene back to the scene containing the main game, effectively resetting the game. When the timer hits zero, the empty scene is selected, which in turn selects the main scene, resetting the timer, score, player and object positions as well as coins.