Load Function
Functionality has been added to the load game button on the main menu. The game is saved upon level completion. If the player were to quit the game and return later, they will be able to press the load game button. Pressing the button will send them to the latest uncompleted level.